requirements.txt
  
gunicorn
flask
flask-cors
SQLAlchemy
Flask-SQLAlchemy
psycopg2
line-bot-sdk
python-dotenv
requests
